CAMEL-23672: TUI - Extract SearchHighlighter and MetricsCollector utilities
Extract two shared utility classes from the TUI to improve maintainability:

- SearchHighlighter: consolidates duplicated find/highlight/wrap logic
  from SourceViewer and LogTab into a single reusable class
- MetricsCollector: moves 30+ ConcurrentHashMap fields and all
  sliding-window update/cleanup/reset logic out of CamelMonitor

CamelMonitor reduced from 3,256 to 2,936 lines. Tab constructors
simplified (EndpointsTab from 10 map params to 1 MetricsCollector).
