+++ dsl/camel-jbang/camel-jbang-plugin-tui/src/main/java/org/apache/camel/dsl/jbang/core/commands/tui/LogTab.java
@@ -38,13 +38,11 @@ import dev.tamboui.text.CharWidth;
 import dev.tamboui.text.Line;
 import dev.tamboui.text.Span;
 import dev.tamboui.text.Text;
-import dev.tamboui.tui.event.KeyCode;
 import dev.tamboui.tui.event.KeyEvent;
 import dev.tamboui.widgets.Clear;
 import dev.tamboui.widgets.block.Block;
 import dev.tamboui.widgets.block.BorderType;
 import dev.tamboui.widgets.block.Title;
-import dev.tamboui.widgets.input.TextInputState;
 import dev.tamboui.widgets.list.ListItem;
 import dev.tamboui.widgets.list.ListState;
 import dev.tamboui.widgets.list.ListWidget;
@@ -94,22 +92,7 @@ class LogTab implements MonitorTab {
     private int hScroll;
     private boolean showLogLevelPopup;
 
-    // Highlight mode: persistent keyword highlighting
-    private String highlightTerm;
-    private Pattern highlightPattern;
-
-    // Find mode: search with next/prev navigation
-    private boolean findInputActive;
-    private boolean highlightInputActive;
-    private TextInputState searchInputState = new TextInputState("");
-    private String findTerm;
-    private Pattern findPattern;
-    private int findMatchIndex = -1;
-    private List<Integer> findMatches = Collections.emptyList();
-
-    private static final Style HIGHLIGHT_STYLE = Style.EMPTY.fg(Color.BLACK).bg(Color.YELLOW);
-    private static final Style FIND_MATCH_STYLE = Style.EMPTY.fg(Color.BLACK).bg(Color.YELLOW);
-    private static final Style FIND_CURRENT_STYLE = Style.EMPTY.fg(Color.BLACK).bg(Color.LIGHT_GREEN);
+    private final SearchHighlighter search = new SearchHighlighter();
 
     LogTab(MonitorContext ctx) {
         this.ctx = ctx;
@@ -132,8 +115,21 @@ class LogTab implements MonitorTab {
 
     @Override
     public boolean handleKeyEvent(KeyEvent ke) {
-        if (findInputActive || highlightInputActive) {
-            return handleSearchInput(ke);
+        if (search.isSearchInputActive()) {
+            boolean handled = search.handleKeyEvent(ke);
+            if (handled && !search.isSearchInputActive() && search.hasFindTerm()) {
+                List<String> plainLines = new ArrayList<>(filteredLogEntries.size());
+                for (LogEntry e : filteredLogEntries) {
+                    plainLines.add(TuiHelper.stripAnsi(e.raw != null ? e.raw : ""));
+                }
+                search.buildFindMatches(plainLines);
+                int newPos = search.jumpToNearestMatch(scroll);
+                if (newPos != scroll) {
+                    followMode = false;
+                    scroll = newPos;
+                }
+            }
+            return handled;
         }
 
         if (showLogLevelPopup) {
@@ -156,22 +152,12 @@ class LogTab implements MonitorTab {
             return true;
         }
 
-        if (ke.isChar('/')) {
-            findInputActive = true;
-            searchInputState = new TextInputState("");
-            return true;
-        }
-        if (ke.isChar('h')) {
-            highlightInputActive = true;
-            searchInputState = new TextInputState("");
-            return true;
-        }
-        if (ke.isChar('n') && findTerm != null) {
-            navigateToNextMatch();
-            return true;
-        }
-        if (ke.isChar('N') && findTerm != null) {
-            navigateToPrevMatch();
+        if (search.handleKeyEvent(ke)) {
+            int matchLine = search.currentMatchLine();
+            if (matchLine >= 0) {
+                followMode = false;
+                scroll = matchLine;
+            }
             return true;
         }
         if (ke.isChar('l') && !ctx.isInfraSelected()) {
@@ -222,59 +208,13 @@ class LogTab implements MonitorTab {
         return false;
     }
 
-    private boolean handleSearchInput(KeyEvent ke) {
-        if (ke.isKey(KeyCode.ESCAPE)) {
-            findInputActive = false;
-            highlightInputActive = false;
-            return true;
-        }
-        if (ke.isConfirm()) {
-            String text = searchInputState.text().trim();
-            if (findInputActive) {
-                if (text.isEmpty()) {
-                    findTerm = null;
-                    findPattern = null;
-                    findMatches = Collections.emptyList();
-                    findMatchIndex = -1;
-                } else {
-                    findTerm = text;
-                    findPattern = Pattern.compile(Pattern.quote(text), Pattern.CASE_INSENSITIVE);
-                    buildFindMatches();
-                    jumpToNearestMatch();
-                }
-                findInputActive = false;
-            } else if (highlightInputActive) {
-                if (text.isEmpty()) {
-                    highlightTerm = null;
-                    highlightPattern = null;
-                } else {
-                    highlightTerm = text;
-                    highlightPattern = Pattern.compile(Pattern.quote(text), Pattern.CASE_INSENSITIVE);
-                }
-                highlightInputActive = false;
-            }
-            return true;
-        }
-        FormHelper.handleTextInput(ke, searchInputState);
-        return true;
-    }
-
     @Override
     public boolean handleEscape() {
-        if (findInputActive || highlightInputActive) {
-            findInputActive = false;
-            highlightInputActive = false;
+        if (search.handleEscape()) {
             return true;
         }
         if (showLogLevelPopup) {
             showLogLevelPopup = false;
-            return true;
-        }
-        if (findTerm != null) {
-            findTerm = null;
-            findPattern = null;
-            findMatches = Collections.emptyList();
-            findMatchIndex = -1;
             return true;
         }
         return false;
@@ -392,21 +332,23 @@ class LogTab implements MonitorTab {
             hScroll = Math.min(hScroll, Math.max(0, cachedLogMaxWidth - visibleWidth));
         }
 
-        if (findPattern != null && entriesChanged) {
-            buildFindMatches();
+        if (search.hasFindTerm() && entriesChanged) {
+            List<String> plainLines = new ArrayList<>(entries.size());
+            for (LogEntry e : entries) {
+                plainLines.add(TuiHelper.stripAnsi(e.raw != null ? e.raw : ""));
+            }
+            search.buildFindMatches(plainLines);
         }
 
         List<Line> allLines = cachedLogLines;
         int start = Math.min(scroll, Math.max(0, allLines.size() - visibleHeight));
         List<Line> visibleLines = allLines.subList(start, Math.min(allLines.size(), start + visibleHeight));
 
-        // Apply highlights only to visible lines
-        if (highlightPattern != null || findPattern != null) {
-            int currentMatchLine = findMatchIndex >= 0 && findMatchIndex < findMatches.size()
-                    ? findMatches.get(findMatchIndex) : -1;
+        int currentMatchLine = search.currentMatchLine();
+        if (currentMatchLine >= 0 || search.hasFindTerm()) {
             List<Line> highlighted = new ArrayList<>(visibleLines.size());
             for (int i = 0; i < visibleLines.size(); i++) {
-                highlighted.add(applyHighlights(visibleLines.get(i), start + i, currentMatchLine));
+                highlighted.add(search.applyHighlights(visibleLines.get(i), start + i, currentMatchLine));
             }
             visibleLines = highlighted;
         }
@@ -436,18 +378,8 @@ class LogTab implements MonitorTab {
 
     @Override
     public void renderFooter(List<Span> spans) {
-        if (findInputActive) {
-            spans.add(Span.styled(" /", HINT_KEY_STYLE));
-            spans.add(Span.raw(searchInputState.text() + "█  "));
-            hint(spans, "Enter", "search");
-            hintLast(spans, "Esc", "cancel");
-            return;
-        }
-        if (highlightInputActive) {
-            spans.add(Span.styled(" h:", HINT_KEY_STYLE));
-            spans.add(Span.raw(searchInputState.text() + "█  "));
-            hint(spans, "Enter", "set");
-            hintLast(spans, "Esc", "cancel");
+        search.renderFooterHints(spans);
+        if (search.isSearchInputActive()) {
             return;
         }
         if (showLogLevelPopup) {
@@ -457,21 +389,13 @@ class LogTab implements MonitorTab {
             return;
         }
 
-        if (findTerm != null) {
-            hint(spans, "Esc", "clear find");
-            hint(spans, "n", "next");
-            hint(spans, "N", "prev");
-            String pos = findMatches.isEmpty()
-                    ? "0/0"
-                    : (findMatchIndex + 1) + "/" + findMatches.size();
-            spans.add(Span.styled("  /", HINT_KEY_STYLE));
-            spans.add(Span.raw("\"" + findTerm + "\" [" + pos + "]  "));
+        if (search.hasFindTerm()) {
+            search.renderFindStatus(spans);
         } else {
             hint(spans, "Esc", "back");
         }
         hint(spans, "↑↓", "scroll");
-        hint(spans, "/", "find");
-        hint(spans, "h", "highlight" + (highlightTerm != null ? " [" + highlightTerm + "]" : ""));
+        search.renderSearchHints(spans);
         hint(spans, "w", "wrap" + (wordWrap ? " [on]" : " [off]"));
         if (!ctx.isInfraSelected()) {
             hint(spans, "l", "level");
@@ -523,128 +447,12 @@ class LogTab implements MonitorTab {
         org.apache.camel.dsl.jbang.core.common.PathUtils.writeTextSafely(root.toJson(), actionFile);
     }
 
-    private void buildFindMatches() {
-        List<Integer> matches = new ArrayList<>();
-        List<LogEntry> entries = filteredLogEntries;
-        for (int i = 0; i < entries.size(); i++) {
-            String plain = TuiHelper.stripAnsi(entries.get(i).raw != null ? entries.get(i).raw : "");
-            if (findPattern.matcher(plain).find()) {
-                matches.add(i);
-            }
-        }
-        findMatches = matches;
-    }
-
-    private void jumpToNearestMatch() {
-        if (findMatches.isEmpty()) {
-            findMatchIndex = -1;
-            return;
-        }
-        for (int i = 0; i < findMatches.size(); i++) {
-            if (findMatches.get(i) >= scroll) {
-                findMatchIndex = i;
-                scrollToMatch();
-                return;
-            }
-        }
-        findMatchIndex = 0;
-        scrollToMatch();
-    }
-
-    private void navigateToNextMatch() {
-        if (findMatches.isEmpty()) {
-            return;
-        }
-        findMatchIndex = (findMatchIndex + 1) % findMatches.size();
-        scrollToMatch();
-    }
-
-    private void navigateToPrevMatch() {
-        if (findMatches.isEmpty()) {
-            return;
-        }
-        findMatchIndex = findMatchIndex <= 0 ? findMatches.size() - 1 : findMatchIndex - 1;
-        scrollToMatch();
-    }
-
-    private void scrollToMatch() {
-        if (findMatchIndex >= 0 && findMatchIndex < findMatches.size()) {
-            followMode = false;
-            scroll = findMatches.get(findMatchIndex);
-        }
-    }
-
     boolean isSearchInputActive() {
-        return findInputActive || highlightInputActive;
+        return search.isSearchInputActive();
     }
 
     void handlePaste(String text) {
-        if (findInputActive || highlightInputActive) {
-            FormHelper.handlePaste(text, searchInputState);
-        }
-    }
-
-    private Line applyHighlights(Line line, int entryIndex, int currentMatchLine) {
-        String fullText = line.rawContent();
-        if (fullText.isEmpty()) {
-            return line;
-        }
-
-        // Collect all match ranges with their styles
-        List<int[]> ranges = new ArrayList<>();
-        List<Style> styles = new ArrayList<>();
-        if (highlightPattern != null) {
-            Matcher m = highlightPattern.matcher(fullText);
-            while (m.find()) {
-                ranges.add(new int[] { m.start(), m.end() });
-                styles.add(HIGHLIGHT_STYLE);
-            }
-        }
-        if (findPattern != null) {
-            boolean isCurrentLine = entryIndex == currentMatchLine;
-            Matcher m = findPattern.matcher(fullText);
-            while (m.find()) {
-                ranges.add(new int[] { m.start(), m.end() });
-                styles.add(isCurrentLine ? FIND_CURRENT_STYLE : FIND_MATCH_STYLE);
-            }
-        }
-        if (ranges.isEmpty()) {
-            return line;
-        }
-
-        // Rebuild spans with highlights applied
-        List<Span> original = line.spans();
-        List<Span> result = new ArrayList<>();
-        int charPos = 0;
-
-        for (Span span : original) {
-            String content = span.content();
-            Style baseStyle = span.style();
-            int spanStart = charPos;
-            int spanEnd = charPos + content.length();
-            int cursor = 0;
-
-            for (int r = 0; r < ranges.size(); r++) {
-                int matchStart = ranges.get(r)[0];
-                int matchEnd = ranges.get(r)[1];
-                if (matchEnd <= spanStart || matchStart >= spanEnd) {
-                    continue;
-                }
-                int localStart = Math.max(0, matchStart - spanStart);
-                int localEnd = Math.min(content.length(), matchEnd - spanStart);
-
-                if (localStart > cursor) {
-                    result.add(Span.styled(content.substring(cursor, localStart), baseStyle));
-                }
-                result.add(Span.styled(content.substring(localStart, localEnd), styles.get(r)));
-                cursor = localEnd;
-            }
-            if (cursor < content.length()) {
-                result.add(Span.styled(content.substring(cursor), baseStyle));
-            }
-            charPos = spanEnd;
-        }
-        return Line.from(result);
+        search.handlePaste(text);
     }
 
     void readNewLogLines(String pid, List<String> newLines) {
