+++ dsl/camel-jbang/camel-jbang-plugin-tui/src/main/java/org/apache/camel/dsl/jbang/core/commands/tui/OverviewTab.java
@@ -77,15 +77,13 @@ class OverviewTab implements MonitorTab {
 
     OverviewTab(
                 MonitorContext ctx,
-                Map<String, LinkedList<Long>> throughputHistory,
-                Map<String, LinkedList<Long>> failedHistory,
-                Map<String, LoadAvg> cpuLoadAvg,
+                MetricsCollector metrics,
                 Set<String> stoppingPids,
                 Runnable onPidChanged) {
         this.ctx = ctx;
-        this.throughputHistory = throughputHistory;
-        this.failedHistory = failedHistory;
-        this.cpuLoadAvg = cpuLoadAvg;
+        this.throughputHistory = metrics.getThroughputHistory();
+        this.failedHistory = metrics.getFailedHistory();
+        this.cpuLoadAvg = metrics.getCpuLoadAvg();
         this.stoppingPids = stoppingPids;
         this.onPidChanged = onPidChanged;
     }
