+++ dsl/camel-jbang/camel-jbang-plugin-tui/src/main/java/org/apache/camel/dsl/jbang/core/commands/tui/SourceViewer.java
@@ -24,8 +24,6 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.IntConsumer;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import dev.tamboui.layout.Constraint;
 import dev.tamboui.layout.Layout;
@@ -41,7 +39,6 @@ import dev.tamboui.tui.event.KeyCode;
 import dev.tamboui.tui.event.KeyEvent;
 import dev.tamboui.widgets.block.Block;
 import dev.tamboui.widgets.block.BorderType;
-import dev.tamboui.widgets.input.TextInputState;
 import dev.tamboui.widgets.paragraph.Paragraph;
 import dev.tamboui.widgets.scrollbar.Scrollbar;
 import dev.tamboui.widgets.scrollbar.ScrollbarState;
@@ -76,23 +73,7 @@ class SourceViewer {
     private IntConsumer onLineSelected;
     private final Map<String, CachedSource> sourceCache = new ConcurrentHashMap<>();
     private boolean wordWrap;
-
-    // Find mode
-    private boolean findInputActive;
-    private boolean highlightInputActive;
-    private TextInputState searchInputState = new TextInputState("");
-    private String findTerm;
-    private Pattern findPattern;
-    private int findMatchIndex = -1;
-    private List<Integer> findMatches = Collections.emptyList();
-
-    // Highlight mode
-    private String highlightTerm;
-    private Pattern highlightPattern;
-
-    private static final Style HIGHLIGHT_STYLE = Style.EMPTY.fg(Color.BLACK).bg(Color.YELLOW);
-    private static final Style FIND_MATCH_STYLE = Style.EMPTY.fg(Color.BLACK).bg(Color.YELLOW);
-    private static final Style FIND_CURRENT_STYLE = Style.EMPTY.fg(Color.BLACK).bg(Color.LIGHT_GREEN);
+    private final SearchHighlighter search = new SearchHighlighter();
 
     private record CachedSource(
             List<String> lines, List<JsonObject> codeData,
@@ -120,14 +101,7 @@ class SourceViewer {
         onLineSelected = null;
         sourceCache.clear();
         wordWrap = false;
-        findInputActive = false;
-        highlightInputActive = false;
-        findTerm = null;
-        findPattern = null;
-        findMatchIndex = -1;
-        findMatches = Collections.emptyList();
-        highlightTerm = null;
-        highlightPattern = null;
+        search.reset();
     }
 
     void setOnLineSelected(IntConsumer callback) {
@@ -138,15 +112,16 @@ class SourceViewer {
         if (!visible) {
             return false;
         }
-        if (findInputActive || highlightInputActive) {
-            return handleSearchInput(ke);
+        if (search.isSearchInputActive()) {
+            boolean handled = search.handleKeyEvent(ke);
+            if (handled && !search.isSearchInputActive() && search.hasFindTerm()) {
+                search.buildFindMatches(lines);
+                selectedLine = search.jumpToNearestMatch(selectedLine);
+            }
+            return handled;
         }
         if (ke.isCancel()) {
-            if (findTerm != null) {
-                findTerm = null;
-                findPattern = null;
-                findMatches = Collections.emptyList();
-                findMatchIndex = -1;
+            if (search.handleEscape()) {
                 return true;
             }
             visible = false;
@@ -158,22 +133,11 @@ class SourceViewer {
             onLineSelected = null;
             return true;
         }
-        if (ke.isChar('/')) {
-            findInputActive = true;
-            searchInputState = new TextInputState("");
-            return true;
-        }
-        if (ke.isChar('h')) {
-            highlightInputActive = true;
-            searchInputState = new TextInputState("");
-            return true;
-        }
-        if (ke.isChar('n') && findTerm != null) {
-            navigateToNextMatch();
-            return true;
-        }
-        if (ke.isChar('N') && findTerm != null) {
-            navigateToPrevMatch();
+        if (search.handleKeyEvent(ke)) {
+            int matchLine = search.currentMatchLine();
+            if (matchLine >= 0) {
+                selectedLine = matchLine;
+            }
             return true;
         }
         if (ke.isChar('w')) {
@@ -224,51 +188,12 @@ class SourceViewer {
         return true;
     }
 
-    private boolean handleSearchInput(KeyEvent ke) {
-        if (ke.isKey(KeyCode.ESCAPE)) {
-            findInputActive = false;
-            highlightInputActive = false;
-            return true;
-        }
-        if (ke.isConfirm()) {
-            String text = searchInputState.text().trim();
-            if (findInputActive) {
-                if (text.isEmpty()) {
-                    findTerm = null;
-                    findPattern = null;
-                    findMatches = Collections.emptyList();
-                    findMatchIndex = -1;
-                } else {
-                    findTerm = text;
-                    findPattern = Pattern.compile(Pattern.quote(text), Pattern.CASE_INSENSITIVE);
-                    buildFindMatches();
-                    jumpToNearestMatch();
-                }
-                findInputActive = false;
-            } else if (highlightInputActive) {
-                if (text.isEmpty()) {
-                    highlightTerm = null;
-                    highlightPattern = null;
-                } else {
-                    highlightTerm = text;
-                    highlightPattern = Pattern.compile(Pattern.quote(text), Pattern.CASE_INSENSITIVE);
-                }
-                highlightInputActive = false;
-            }
-            return true;
-        }
-        FormHelper.handleTextInput(ke, searchInputState);
-        return true;
-    }
-
     boolean isSearchInputActive() {
-        return findInputActive || highlightInputActive;
+        return search.isSearchInputActive();
     }
 
     void handlePaste(String text) {
-        if (findInputActive || highlightInputActive) {
-            FormHelper.handlePaste(text, searchInputState);
-        }
+        search.handlePaste(text);
     }
 
     void render(Frame frame, Rect area) {
@@ -312,8 +237,7 @@ class SourceViewer {
             scrollX = Math.min(scrollX, maxHScroll);
         }
 
-        int currentMatchLine = findMatchIndex >= 0 && findMatchIndex < findMatches.size()
-                ? findMatches.get(findMatchIndex) : -1;
+        int currentMatchLine = search.currentMatchLine();
 
         int end = Math.min(scrollY + visibleLines, lines.size());
         List<Line> visible = new ArrayList<>();
@@ -321,9 +245,7 @@ class SourceViewer {
             String raw = lines.get(i);
             boolean isSelected = (i == selectedLine);
             Line line = highlightSourceLine(raw, hSkip, isSelected, inner.width());
-            if (highlightPattern != null || findPattern != null) {
-                line = applySearchHighlights(line, i, currentMatchLine);
-            }
+            line = search.applyHighlights(line, i, currentMatchLine);
             visible.add(line);
         }
 
@@ -350,35 +272,17 @@ class SourceViewer {
     }
 
     void renderFooter(List<Span> spans) {
-        if (findInputActive) {
-            spans.add(Span.styled(" /", MonitorContext.HINT_KEY_STYLE));
-            spans.add(Span.raw(searchInputState.text() + "█  "));
-            MonitorContext.hint(spans, "Enter", "search");
-            MonitorContext.hintLast(spans, "Esc", "cancel");
+        search.renderFooterHints(spans);
+        if (search.isSearchInputActive()) {
             return;
         }
-        if (highlightInputActive) {
-            spans.add(Span.styled(" h:", MonitorContext.HINT_KEY_STYLE));
-            spans.add(Span.raw(searchInputState.text() + "█  "));
-            MonitorContext.hint(spans, "Enter", "set");
-            MonitorContext.hintLast(spans, "Esc", "cancel");
-            return;
-        }
-        if (findTerm != null) {
-            MonitorContext.hint(spans, "Esc", "clear find");
-            MonitorContext.hint(spans, "n", "next");
-            MonitorContext.hint(spans, "N", "prev");
-            String pos = findMatches.isEmpty()
-                    ? "0/0"
-                    : (findMatchIndex + 1) + "/" + findMatches.size();
-            spans.add(Span.styled("  /", MonitorContext.HINT_KEY_STYLE));
-            spans.add(Span.raw("\"" + findTerm + "\" [" + pos + "]  "));
+        if (search.hasFindTerm()) {
+            search.renderFindStatus(spans);
         } else {
             MonitorContext.hint(spans, "Esc/c", "close");
         }
         MonitorContext.hint(spans, "↑↓", "navigate");
-        MonitorContext.hint(spans, "/", "find");
-        MonitorContext.hint(spans, "h", "highlight" + (highlightTerm != null ? " [" + highlightTerm + "]" : ""));
+        search.renderSearchHints(spans);
         MonitorContext.hint(spans, "w", "wrap" + (wordWrap ? " [on]" : " [off]"));
         if (!wordWrap) {
             MonitorContext.hint(spans, "←→", "horizontal");
@@ -669,112 +573,6 @@ class SourceViewer {
         }
 
         return full;
-    }
-
-    private void buildFindMatches() {
-        List<Integer> matches = new ArrayList<>();
-        for (int i = 0; i < lines.size(); i++) {
-            if (findPattern.matcher(lines.get(i)).find()) {
-                matches.add(i);
-            }
-        }
-        findMatches = matches;
-    }
-
-    private void jumpToNearestMatch() {
-        if (findMatches.isEmpty()) {
-            findMatchIndex = -1;
-            return;
-        }
-        for (int i = 0; i < findMatches.size(); i++) {
-            if (findMatches.get(i) >= selectedLine) {
-                findMatchIndex = i;
-                scrollToMatch();
-                return;
-            }
-        }
-        findMatchIndex = 0;
-        scrollToMatch();
-    }
-
-    private void navigateToNextMatch() {
-        if (findMatches.isEmpty()) {
-            return;
-        }
-        findMatchIndex = (findMatchIndex + 1) % findMatches.size();
-        scrollToMatch();
-    }
-
-    private void navigateToPrevMatch() {
-        if (findMatches.isEmpty()) {
-            return;
-        }
-        findMatchIndex = findMatchIndex <= 0 ? findMatches.size() - 1 : findMatchIndex - 1;
-        scrollToMatch();
-    }
-
-    private void scrollToMatch() {
-        if (findMatchIndex >= 0 && findMatchIndex < findMatches.size()) {
-            selectedLine = findMatches.get(findMatchIndex);
-        }
-    }
-
-    private Line applySearchHighlights(Line line, int lineIndex, int currentMatchLine) {
-        String fullText = line.rawContent();
-        if (fullText.isEmpty()) {
-            return line;
-        }
-
-        List<int[]> ranges = new ArrayList<>();
-        List<Style> rangeStyles = new ArrayList<>();
-        if (highlightPattern != null) {
-            Matcher m = highlightPattern.matcher(fullText);
-            while (m.find()) {
-                ranges.add(new int[] { m.start(), m.end() });
-                rangeStyles.add(HIGHLIGHT_STYLE);
-            }
-        }
-        if (findPattern != null) {
-            boolean isCurrentLine = lineIndex == currentMatchLine;
-            Matcher m = findPattern.matcher(fullText);
-            while (m.find()) {
-                ranges.add(new int[] { m.start(), m.end() });
-                rangeStyles.add(isCurrentLine ? FIND_CURRENT_STYLE : FIND_MATCH_STYLE);
-            }
-        }
-        if (ranges.isEmpty()) {
-            return line;
-        }
-
-        List<Span> original = line.spans();
-        List<Span> result = new ArrayList<>();
-        int charPos = 0;
-        for (Span span : original) {
-            String content = span.content();
-            Style baseStyle = span.style();
-            int spanStart = charPos;
-            int spanEnd = charPos + content.length();
-            int cursor = 0;
-            for (int r = 0; r < ranges.size(); r++) {
-                int matchStart = ranges.get(r)[0];
-                int matchEnd = ranges.get(r)[1];
-                if (matchEnd <= spanStart || matchStart >= spanEnd) {
-                    continue;
-                }
-                int localStart = Math.max(0, matchStart - spanStart);
-                int localEnd = Math.min(content.length(), matchEnd - spanStart);
-                if (localStart > cursor) {
-                    result.add(Span.styled(content.substring(cursor, localStart), baseStyle));
-                }
-                result.add(Span.styled(content.substring(localStart, localEnd), rangeStyles.get(r)));
-                cursor = localEnd;
-            }
-            if (cursor < content.length()) {
-                result.add(Span.styled(content.substring(cursor), baseStyle));
-            }
-            charPos = spanEnd;
-        }
-        return Line.from(result);
     }
 
     static int findLicenseHeaderEnd(List<JsonObject> codeLines) {
